requirements.txt
Flask>=3.0.0
openai>=1.0.0
python-dotenv
